+++ packmol_step/installer.py
@@ -54,40 +54,183 @@ class Installer(seamm_installer.InstallerBase):
         logger.debug("Initializing the PACKMOL installer object.")
 
         self.section = "packmol-step"
-        self.path_name = "packmol-path"
         self.executables = ["packmol"]
         self.resource_path = Path(pkg_resources.resource_filename(__name__, "data/"))
+
+        # The environment.yaml file for Conda installations.
+        logger.debug(f"data directory: {self.resource_path}")
+        self.environment_file = self.resource_path / "seamm-packmol.yml"
+
+    def check(self):
+        """Check the status of the Packmol installation."""
+        print("Checking the Packmol installation.")
+
         # What Conda environment is the default?
-        data = self.configuration.get_values(self.section)
+        path = self.configuration.path.parent / "packmol.ini"
+        if not path.exists():
+            text = (self.resource_path / "packmol.ini").read_text()
+            path.write_text(text)
+            print(f"    The packmol.ini file did not exist. Created {path}")
+
+        self.exe_config.path = path
+
+        # Get the current values
+        data = self.exe_config.get_values("local")
+
         if "conda-environment" in data and data["conda-environment"] != "":
             self.environment = data["conda-environment"]
         else:
             self.environment = "seamm-packmol"
 
-        # The environment.yaml file for Conda installations.
-        path = Path(pkg_resources.resource_filename(__name__, "data/"))
-        logger.debug(f"data directory: {path}")
-        self.environment_file = path / "seamm-packmol.yml"
+        super().check()
 
-    def exe_version(self, path):
+    def install(self):
+        """Install Packmol in a conda environment."""
+        print("Installing Packmol.")
+
+        # What Conda environment is the default?
+        path = self.configuration.path.parent / "packmol.ini"
+        if not path.exists():
+            text = (self.resource_path / "packmol.ini").read_text()
+            path.write_text(text)
+            print(f"    The packmol.ini file did not exist. Created {path}")
+
+        self.exe_config.path = path
+
+        # Get the current values
+        data = self.exe_config.get_values("local")
+
+        if "conda-environment" in data and data["conda-environment"] != "":
+            self.environment = data["conda-environment"]
+        else:
+            self.environment = "seamm-packmol"
+
+        super().install()
+
+    def show(self):
+        """Show the status of the Packmol installation."""
+        print("Showing the Packmol installation.")
+
+        # What Conda environment is the default?
+        path = self.configuration.path.parent / "packmol.ini"
+        if not path.exists():
+            text = (self.resource_path / "packmol.ini").read_text()
+            path.write_text(text)
+            print(f"    The packmol.ini file does not exist at {path}")
+            print("    The 'check' command will create it if Packmol is installed.")
+            print("    Otherwise 'install' will install Packmol.")
+            return
+
+        self.exe_config.path = path
+
+        if not self.exe_config.section_exists("local"):
+            print(
+                "    Packmol is not configured: there is no 'local' section in "
+                f"     {path}."
+            )
+            return
+
+        # Get the current values
+        data = self.exe_config.get_values("local")
+
+        if "conda-environment" in data and data["conda-environment"] != "":
+            self.environment = data["conda-environment"]
+        else:
+            self.environment = "seamm-packmol"
+
+        super().show()
+
+    def uninstall(self):
+        """Uninstall the Packmol installation."""
+        print("Uninstall the Packmol installation.")
+
+        # What Conda environment is the default?
+        path = self.configuration.path.parent / "packmol.ini"
+        if not path.exists():
+            text = (self.resource_path / "packmol.ini").read_text()
+            path.write_text(text)
+            print(
+                f""""    The packmol.ini file does not exist at {path}
+    Perhaps Packmol is not installed, but if it is the 'check' command may locate it
+    and create the ini file, after which 'uninstall' will remove it."""
+            )
+            return
+
+        self.exe_config.path = path
+
+        if not self.exe_config.section_exists("local"):
+            print(
+                f""""    The packmol.ini file at {path} does not have local section.
+    Perhaps Packmol is not installed, but if it is the 'check' command may locate it
+    and update the ini file, after which 'uninstall' will remove it."""
+            )
+            return
+
+        # Get the current values
+        data = self.exe_config.get_values("local")
+
+        if "conda-environment" in data and data["conda-environment"] != "":
+            self.environment = data["conda-environment"]
+        else:
+            self.environment = "seamm-packmol"
+
+        super().uninstall()
+
+    def update(self):
+        """Updates the Packmol installation."""
+        print("Updating the Packmol installation.")
+
+        # What Conda environment is the default?
+        path = self.configuration.path.parent / "packmol.ini"
+        if not path.exists():
+            text = (self.resource_path / "packmol.ini").read_text()
+            path.write_text(text)
+            print(f"    The packmol.ini file did not exist. Created {path}")
+
+        self.exe_config.path = path
+
+        # Get the current values
+        data = self.exe_config.get_values("local")
+
+        if "conda-environment" in data and data["conda-environment"] != "":
+            self.environment = data["conda-environment"]
+        else:
+            self.environment = "seamm-packmol"
+
+        super().update()
+
+    def exe_version(self, config):
         """Get the version of the PACKMOL executable.
 
         Parameters
         ----------
-        path : pathlib.Path
-            Path to the executable.
+        config : dict
+            Dictionary of options for running Packmol
 
         Returns
         -------
-        str
+        "Packmol", str
             The version reported by the executable, or 'unknown'.
         """
+        environment = config["conda-environment"]
+        conda = config["conda"]
+        if environment[0] == "~":
+            environment = str(Path(environment).expanduser())
+            command = f"'{conda}' run --live-stream -p '{environment}'"
+        elif Path(environment).is_absolute():
+            command = f"'{conda}' run --live-stream -p '{environment}'"
+        else:
+            command = f"'{conda}' run --live-stream -n '{environment}'"
+        command += " packmol -log none"
+
+        logger.debug(f"    Running {command}")
         try:
             result = subprocess.run(
-                [str(path), "-log", "none"],
+                command,
                 stdin=subprocess.DEVNULL,
                 capture_output=True,
                 text=True,
+                shell=True,
             )
         except Exception:
             version = "unknown"
@@ -103,4 +246,4 @@ class Installer(seamm_installer.InstallerBase):
                         version = value
                         break
 
-        return version
+        return "Packmol", version