+++ HISTORY.rst
@@ -5,6 +5,11 @@ History
     * Using RDKit for SMILES since we found some issues with OpenBabel, and also the
       atom typing uses RDKit, so this is more compatible.
       
+2024.3.19 -- Updated installer for new scheme
+    * packmol-step-installer now uses the new scheme, which supports both Conda and
+      Docker installation.
+    * Added seamm-packmol Docker image
+
 2024.1.16 -- Adding support for containers
     * Added the ability to work in Docker containers.